+++ fastapi-solution/src/models/genre.py
@@ -1,13 +1,6 @@
-import orjson
-from models.functions_for_models import orjson_dumps
-from pydantic import BaseModel
+from models.base_model import BasicModel
 
 
-class Genre(BaseModel):
+class Genre(BasicModel):
     id: str
     genre_name: str
-
-    class Config:
-        # Заменяем стандартную работу с json на более быструю.
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
