+++ fastapi-solution/src/models/person.py
@@ -1,14 +1,7 @@
-import orjson
-from models.functions_for_models import orjson_dumps
-from pydantic import BaseModel
+from models.base_model import BasicModel
 
 
-class Person(BaseModel):
+class Person(BasicModel):
     id: str
     film_work: list = []
     role: list = []
-
-    class Config:
-        # Заменяем стандартную работу с json на более быструю.
-        json_loads = orjson.loads
-        json_dumps = orjson_dumps
